Extract inline JS to appstatic and fix dastardly ID bug
Turns out that in HTML4, you can't have IDs that start with integers... but sometimes you can. In Chrome this never surfaced, but in Firefox it means that weirdly, some swatches can't be selected when modifying a collection. Adding a letter to the beginning of every card ID fixes this issue and makes them all select appropriately.